Patch 0.1.2: Updated subdirectories
