+++ src/components/FadingTextCycle.js
