+++ src/components/RepositoryList.js
@@ -1,8 +1,7 @@
 import React from "react";
 import Grid from "@material-ui/core/Grid";
 import { motion } from "framer-motion";
-import RepositoryCard from "./RepositoryCard";
-import MainRepositoryCard from "./MainRepositoryCard";
+import RepositoryCard from "./subcomponents/RepositoryCard";
 import Card from "@material-ui/core/Card";
 
 const transition = {
@@ -14,15 +13,9 @@ const projectVariants = {
     exit: { opacity: 0, transition },
 };
 
-function SmallRepositoryList(props) {
+function RepositoryList(props) {
     const { repos, xsNum, mdNum } = props;
     const noRepoMessage = 'No repositories loaded';
-    const mainProjectsArray = [
-        'DnDLevelGen',
-        'minericpark.github.io-redesign',
-        'HitBeat',
-        'IMDbSearch'
-    ];
 
     if (!repos || repos.length === 0) {
         return (<RepositoryCard title={noRepoMessage}/>);
@@ -30,24 +23,23 @@ function SmallRepositoryList(props) {
 
     return (
         <motion.div variants={{ enter: { transition: { staggerChildren: 0.3 } } }}>
-            <Grid container spacing={6}>
+            <Grid container spacing={4}>
                 {repos.filter(repo => repo.fork === false)
                     .filter(repo => repo.language != null)
                     .sort(function(a, b) {
                         return b.updated_at.localeCompare(a.updated_at);
                     })
-                    .filter(repo => mainProjectsArray.indexOf(repo.name) >= 0)
                     .map((repo) => (
                         <Grid item xs={xsNum} md={mdNum}>
                             <Card style={{height: '100%'}} component={motion.div} variants={projectVariants} whileHover={{ scale: 1.1 }}>
-                                <MainRepositoryCard title={repo.name} subTitle={repo.language} mainText={repo.description} href={repo.html_url}/>
+                                <RepositoryCard title={repo.full_name} subTitle={repo.language} mainText={repo.description} href={repo.html_url}/>
                             </Card>
                         </Grid>
-                    ))}
+                ))}
             </Grid>
         </motion.div>
     );
 
 }
 
-export default SmallRepositoryList;
+export default RepositoryList;
