+++ src/pages/blocks/BlogBlock.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Grid from "@material-ui/core/Grid";
-import BlogPosts from "../../components/subcomponents/BlogPosts";
-import Sidebar from "../../components/subcomponents/Sidebar";
+import BlogPosts from "../../components/BlogPosts";
+import Sidebar from "../../components/Sidebar";
 import {makeStyles} from "@material-ui/core/styles";
 import PropTypes from "prop-types";
 import {motion} from "framer-motion";
