+++ src/pages/blocks/ProfessionBlock.js
@@ -1,9 +1,9 @@
 import Grid from "@material-ui/core/Grid";
 import Typography from "@material-ui/core/Typography";
 import Divider from "@material-ui/core/Divider";
-import SpringImageCard from "../../components/subcomponents/SpringImageCard";
-import ImageCard from "../../components/subcomponents/ImageCard";
-import PdfViewer from "../../components/subcomponents/PdfViewer";
+import SpringImageCard from "../../components/SpringImageCard";
+import ImageCard from "../../components/ImageCard";
+import PdfViewer from "../../components/PdfViewer";
 import React from "react";
 import {makeStyles} from "@material-ui/core/styles";
 import {motion} from "framer-motion";
